requirements.txt
# Backend (GraphRAG core) — DO NOT CHANGE
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.1
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
pdfplumber>=0.9.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pydantic>=2.4.0

# Web server (replaces Streamlit)
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9